api/requirements.txt
django==5.1
django-ninja==1.2.2
psycopg==3.2.1
django-ninja-jwt==5.3.3